add types for header megamenu
