+++ src/templates/globals/wrapper/index.tsx
@@ -4,10 +4,14 @@ import { Banner } from '@/templates/globals/banners/banner'
 import { PromoBanner } from '@/templates/globals/banners/promoBanner'
 import { FacilityBanner } from '@/templates/globals/banners/facilityBanner'
 import { isEmpty } from 'lodash'
-import { BannerType, FacilityBannerType, PromoBannerType } from '@/types/index'
+import {
+  BannerType,
+  FacilityBannerType,
+  PromoBannerType,
+  HeaderFooterData,
+} from '@/types/index'
 import { NodeBanner } from '@/types/dataTypes/drupal/node'
 import { BannerDisplayType, BannerTypeMapping } from '@/data/queries/banners'
-import { HeaderFooterData } from '@/data/queries/headerFooter'
 import { Header } from '../header'
 import { Footer } from '../footer/index'
